+++ src/sm_tracker/platforms/__init__.py
@@ -44,6 +44,8 @@ SUPPORTED_PLATFORM_NAMES: tuple[str, ...] = (
     "mastodon",
     "threads",
     "twitter",
+    "instagram",
+    "youtube",
 )
 
 
@@ -54,9 +56,11 @@ def resolve_adapters(
     from sm_tracker.platforms.bluesky import create_bluesky_adapter
     from sm_tracker.platforms.facebook import create_facebook_adapter
     from sm_tracker.platforms.farcaster import create_farcaster_adapter
+    from sm_tracker.platforms.instagram import create_instagram_adapter
     from sm_tracker.platforms.mastodon import create_mastodon_adapter
     from sm_tracker.platforms.threads import create_threads_adapter
     from sm_tracker.platforms.twitter import create_twitter_adapter
+    from sm_tracker.platforms.youtube import create_youtube_adapter
 
     env_map = os.environ if env is None else env
     factories: dict[str, Callable[[Mapping[str, str]], PlatformAdapter]] = {
@@ -66,6 +70,8 @@ def resolve_adapters(
         "mastodon": create_mastodon_adapter,
         "threads": create_threads_adapter,
         "twitter": create_twitter_adapter,
+        "instagram": create_instagram_adapter,
+        "youtube": create_youtube_adapter,
     }
     adapters: list[PlatformAdapter] = []
     warnings: list[str] = []